requirements.txt
bcrypt==3.2.2
bidict==0.22.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
django-environ==0.9.0
dnspython==1.16.0
eventlet==0.30.2
Flask==2.1.3
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-engineio==4.3.3
python-socketio==5.7.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1